Remove shortened field variants
These were initially added to provide decorators with the ability to
opt-out of name cropping, however as there's not actually a real use
case for that as-yet we'll just remove them. The 'name' and 'fullTitle'
fields will now contain the cropped strings, which simplifies the life
of filter hooks as they don't need to set both the full-length and
shortened versions of fields if modifying either.