Remove nameplates core from module system
The nameplates core module itself was largely a misfeature in that it
gave users a fairly nasty footgun where nameplate decorations wouldn't
work unless *both* the core module and an associated decorator were
simultaneously enabled.

As such the ability to disable the core nameplates module has been
removed, in favour of just allowing users to disable decorators for
the same effect.

The core module has been migrated to an Ace3 module and internally
is disabled by default - it will enable itself upon registration of
the first "OnNamePlateDataUpdated" callback as it does today.

As an extension - the core now also supports runtime disabling; if all
decorators unregister their "OnNamePlateDataUpdated" callbacks then
the core will disable itself and nameplates will no longer be decorated.

Expanding runtime disabling to the Blizzard/Kui decorators isn't
going to be done as part of this PR, but this does lay the groundwork
for that to be added in the future. As a test however, one can simply
"/run TRP3_NamePlates:Disable()" and "/run TRP3_NamePlates:Enable()"
to verify things work as intended.