Remove TRP3_NAMEPLATES_ADDON global
This global was used to acquire an exclusive lock on nameplate
decoration rights, however its usage was reasonably complicated and
there wasn't much gain - the odds of someone having two eligible
nameplate addons installed is fairly low to begin with.

As such we'll just let it be a free-for-all. If a user has multiple
nameplate addons installed the decorators should be self-contained
anyway and not conflict.