Improve target visibility after loading screens
The PLAYER_TARGET_CHANGED event on a UI reload will fire before any
nameplates are set up - which leads to a small bug where if the
user has a target selected, reloads their UI, and then changes
target the previous unit won't have their nameplate hidden if
configured to do so.

We'll update the target token for nameplates each time the nameplate
associated with the target unit changes. This should improve the
reliability somewhat.