Rework logging infrastructure
This commit implements a much simpler logging infrastructure and
migrates all existing usages of both old logging systems into the
new one.

The logging API surface has been simplified to two function calls:

  - TRP3_API.Log(message, ...)
    Joins all supplied arguments with a " " delimiter and writes the
	message.

  - TRP3_API.Logf(format, ...)
    Formats a log message and then writes it out.

Both of these functions cannot error; if an error occurs either when
converting parameters to strings, when formatting the message, or
when submitting it to a log sink that error will be forwarded to the
global error handler but will not break any calling code.

As with our existing logging implementation all log messages will be
forwarded to the following sinks:

  - Any chat frame named "Logs" if debug mode is enabled.
  - The DebugLog addon if it is present.

As an addition we now store the most recent 1024 messages in a circular
buffer. We could look into augmenting error reports with this info
one day, or providing users with a way to copy/paste recent messages
for support requests.

The concept of log levels has been removed. We never actually provided
any means for users (or ourselves) to control the verbosity of logging,
nor implemented a logging frame - as well as *most* of our messages not
specifying a level and defaulting to "INFO".

If something is worth logging then it probably shouldn't ever be
filtered anyway, so there's not much point in keeping that concept
around.