NPC nameplate support and additional settings (#702)
* Reduce function calls for nameplate settings

Each nameplate settings lookup involved about three or four function
calls plus string junk being generated by an assertion in the
getValue function. This would be repeated for each setting, and
potentially for all on-screen nameplates.

We now cache nameplate settings in a new global table and read from
that directly. The table will be re-loaded on configuration changes
at runtime.

* Add hooks for customizing NPC nameplates

NPC unit nameplates are now customized and will by default respect
the "Hide non-roleplay units" option. Customization of NPC nameplates
can be disabled through a new "Disable customizations on NPC units"
option, which is off by default.

A new display info filter mechanism has been added permit modifying the
nameplate display tables prior to them being submitted to decorators;
this can be used by Extended to customize NPC nameplates further if
desired.

* Add settings for name/title cropping

* Small cleanup of Kui handlers

* Disable crop sliders if names/titles are disabled

* Add some missing lamb sauce

* Pre-crop names and titles in display info setup

Rather than having the decorators always manage cropping themselves,
two new fields (shortenedName and shortenedFullTitle) have been added
to display info which contain the cropped equivalents of the name and
title strings. Use of these is optional but recommended.

* Group cropping sliders closer to other settings

* Small clarification for titles in name cropping

Short titles get prepended to the name and will count toward the
cropping size limits.

* Add option for displaying first names only

As an alternative to cropping long name texts, it makes sense to add
an option to also just display first names only. This can only be
applied to player profiles and won't work with anything received
via MSP.

* Small help text tweaks

* Add utility function for cropping names/titles

This can be used in-place of the shortenedName/shortenedFullTitle fields
if desired, or when replacing the contents of this fields in filters
without needing to know about the TRP3_NamePlatesSettings table.

* Remove shortened field variants

These were initially added to provide decorators with the ability to
opt-out of name cropping, however as there's not actually a real use
case for that as-yet we'll just remove them. The 'name' and 'fullTitle'
fields will now contain the cropped strings, which simplifies the life
of filter hooks as they don't need to set both the full-length and
shortened versions of fields if modifying either.

* Remove ticker protection from request queue

As of 10.0.0/3.4.1 any errors in tickers don't prevent later executions
of the ticker from ever happening. This isn't the case in Classic Era
yet, but this code hasn't errored a single time since deployment
anyway so the protections probably aren't needed.

* Remove nameplates core from module system

The nameplates core module itself was largely a misfeature in that it
gave users a fairly nasty footgun where nameplate decorations wouldn't
work unless *both* the core module and an associated decorator were
simultaneously enabled.

As such the ability to disable the core nameplates module has been
removed, in favour of just allowing users to disable decorators for
the same effect.

The core module has been migrated to an Ace3 module and internally
is disabled by default - it will enable itself upon registration of
the first "OnNamePlateDataUpdated" callback as it does today.

As an extension - the core now also supports runtime disabling; if all
decorators unregister their "OnNamePlateDataUpdated" callbacks then
the core will disable itself and nameplates will no longer be decorated.

Expanding runtime disabling to the Blizzard/Kui decorators isn't
going to be done as part of this PR, but this does lay the groundwork
for that to be added in the future. As a test however, one can simply
"/run TRP3_NamePlates:Disable()" and "/run TRP3_NamePlates:Enable()"
to verify things work as intended.

* Names aren't titles silly

* Remove TRP3_NAMEPLATES_ADDON global

This global was used to acquire an exclusive lock on nameplate
decoration rights, however its usage was reasonably complicated and
there wasn't much gain - the odds of someone having two eligible
nameplate addons installed is fairly low to begin with.

As such we'll just let it be a free-for-all. If a user has multiple
nameplate addons installed the decorators should be self-contained
anyway and not conflict.

* Mark Kui dependencies as externals

* Add option to disable nameplates in instances

This is primarily intended for the NPC nameplate customization that
Extended may offer; for friendly nameplates disabling this setting will
have no impact, but as we're now permitting NPC nameplates to be
hidden this would have an impact while in instanced content.

This option is enabled by default.

* Initialize in-combat status on module enablement

This is to deal with any edge cases if the nameplate core gets
disabled while in one combat state, and re-enabled in another.

* Add "Always show target unit" option

This option - enabled by default - will override the other "Hide..."
options for the nameplate of the players' current target. Odds are
if the player has something targetted the nameplate would be of
actual interest to see, so having this enabled by default makes sense.

* Less words

* Improve target visibility after loading screens

The PLAYER_TARGET_CHANGED event on a UI reload will fire before any
nameplates are set up - which leads to a small bug where if the
user has a target selected, reloads their UI, and then changes
target the previous unit won't have their nameplate hidden if
configured to do so.

We'll update the target token for nameplates each time the nameplate
associated with the target unit changes. This should improve the
reliability somewhat.