Add guild name/rank misc. info presets (#704)
* Add guild name/rank misc. info presets

Two new presets have been added for guild name and rank to the
additional info fields in a profile. Like pronouns, these fields are
treated specially for both display in tooltips and via MSP comms.

When showing a tooltip for a player if they have a custom guild name
set then it will fully override and replace the guild name that would
have otherwise been shown based off their players' guild affiliation.
The rank will be set to that of the "Guild rank" info field if present,
or will default to a sensible "Member" string.

If the player has only a custom guild rank but no guild name, then only
the rank of their players' actual guild will be replaced.

For MSP comms two new fields have been added for guild name and rank;
"PG" and "PR" respectively. Use of the "G*" namespace was avoided due
to that being used for "game info" fields like unit GUID.

* Add numeric ID field to Misc. info presets

Our existing misc info field system has some annoying complexity where
multiple parts of the code (tooltips and MSP) need to check for both
localized and english names of fields when looking for "special"
things like pronouns.

This has a few issues - most notably that people who send a non-english
field name to someone else running in a different language won't
result in the field displaying correctly in the tooltip.

With these changes misc. info fields now have an integral "ID" field
that identifies the source type (or preset) a field was initially
created from.

Note that it is possible for users to still rename fields created
from presets - so you *can* run into weird situations if you've
created a "Pronouns" field and then renamed it to "Guild name"; the
ID would still register it as Pronouns.

Despite this however, that one downside is far less problematic than
the localization problem - so we can live with it for now. Some
new utilities have been added to work with misc. types and to obtain
structured views of misc. info fields that don't expose the inner
layout of the saved data.

* Expose custom guild data via nameplates

* Rename physiognomy to "Facial features"

"Facial" now looks like a bit of a weird word.

* Fix small typo

* Add voice reference preset and tooltip display

* Don't overwrite pronouns silly

* Add IDs to imported MI fields