Add option to disable welcome message
Some users prefer not getting spammed by fifty addons announcing their
existence on login, so to appease these people this adds a setting
to at least disable part of the welcome message printed on login.

Unfortunately the "The profile <name> is loaded" message isn't
covered by this, but that can be solved later.