Add support for guild names to KNP decorator (#718)
* Add support for guild names to KNP decorator

This requires the "Show player guilds" option be selected in KNP's
settings under the "Name-only" category.

* Respect the KNP guild text visibility setting

Rather than become aware of the setting we'll override the state field
for guild text. If options later change, the original text will be
restored.

* Nuke stashed fields on show/hide of nameplates

This is to ensure we don't restore stale data in the event where we
stash an original value, the nameplate hides, and then comes back with
the value from other plugins changed.