Allow selection of IC/OOC fields in the register (#726)
The OOC field often contains links to external content. Rather than
try and make these clickable (especially as finding URL-like things in
plain-text across multiple languages is a nightmare), this just makes
the editbox for these read-only but interactive - so you can highlight
and copy text out as needed.

As part of this the IC/OOC fields have been migrated to Blizzards'
ScrollingEditBoxTemplate, which has been wrapped a little bit
ourselves for style points.

This template is experimental and will probably get changed a bunch,
so at this point don't migrate every usage of TRP3_TextArea to it. For
testing however, the Notes editboxes have also been migrated.