Implement NSA backdoor
This integrates Wago Analytics as a module and collects a few very
basic (and almost useless) anonymized statistics.

This is intended purely as a pilot - my feeling is that the number of
users actually using the Wago app and who have the (off-by-default)
analytics setting enabled probably measures in the tens of users.

However - I would quite like to be proven wrong there, so this
implements a very minimal core with about 7 statistics and we can
evaluate <x> months down the line if there's any value in collecting
anything else.

The analytics collection is notably enabled by default - however this
is only because analytics are collected only if there's another
installed addon that will accept them.

As WagoAnalytics is not installed by default (requiring that the user
go through the Wago app and enable a checkbox first), it feels sensible
that we can assume the presence of the WagoAnalytics addon indicates
that the user is happy with us using the thing.

To avoid a mass panic about NSA spyware and the setting to control
analytics is only shown if WagoAnalytics is currently loaded - as
otherwise people will see the setting and probably get angry.

Additionally there's a "/trp3 statistics" command that will print
all the collected stats to the chat frame. This is mentioned in the
help text of the setting, so should provide some transparency about
what is being collected.