Add pre-commit configuration
This sets up a minimal pre-commit configuration to apply common checks
that largely match what our existing actions were already checking.

Because luarocks is complete and utter garbage the configuration
elects to not use the upstream hook that luacheck provides, instead
favouring a system integration - so you'll need luacheck on your
path for this to work.

Likewise for XML validation you'll need xmllint. For Windows this
is a bit tricky; the easiest solution I've come up with here is
just going to libxml2's GitLab instance and grabbing the artefacts
of one of their Windows builds:

  https://gitlab.gnome.org/GNOME/libxml2/-/jobs

Use of pre-commit isn't required but should hopefully avoid those
arguments we keep having with Bob.