Move Plater mod logic to TRP3
In the past, the Plater mod itself held all of the functions and everything necessary to customize nameplates. Moving this to TRP3 will allow more visibility into the inner workings of the mod, and allow Meo to find all the problems for me.

The only downside to this approach is that it means the Plater mod now depends on the TRP module to work at all, which, strangely, the old version could work without it.