Tweaks to nameplate visibility rules
The "Always show target" option had a rather confusing effect of not
applying if the unit was OOC and "Hide out of character units" was
selected.

As it feels like this setting should be a stronger override, the
"Always show target" option will now indeed do so - and will always
show nameplates of target units irrespective of other "Hide ..."
options.

As part of this, nameplate visibility rules have been combined into a
single evaluation function to ensure consistent application.