Update Plater mod and move nameplate customization to TRP3 (#741)
* Move Plater mod logic to TRP3

In the past, the Plater mod itself held all of the functions and everything necessary to customize nameplates. Moving this to TRP3 will allow more visibility into the inner workings of the mod, and allow Meo to find all the problems for me.

The only downside to this approach is that it means the Plater mod now depends on the TRP module to work at all, which, strangely, the old version could work without it.

* Use the event like a normal person

* Fixing everything

:100: :100: :100: :100: :100:

* Adding definitely not-forgotten fix

I forgot to change the Plater mod name bad from the dev version. :)))

* "It's Plater"

This fixes the nameplate getting reverted to default. Uses the Plater "Nameplate Added" hook to refresh nameplates now as well.

* Sanity check

* Only hide guild name if it's explicitly disabled

* Yeet A

don't even ask

* Quell Luacheck's antics

* Check for nil displayInfo earlier

* Rename RPDisplayName variable for consistency

* Simplify unit frame visibility

* Lamb sauce

* Misc. stylistic fixups

* Take a break from the scripts

There shouldn't ever be more than one of these scripts since it overwrites it by name.

Co-authored-by: Daniel Yates <me@meorawr.io>

---------

Co-authored-by: Daniel Yates <me@meorawr.io>