Default guild rank text on display
If the user puts in a blank or space-only string to force the guild
rank to be hidden, we'll default it to "Member" in the tooltip.

Hiding the rank information in this way doesn't really work too well
in contexts that isn't the tooltip; for example rank information isn't
shown in nameplates so the standalone guild name string can look quite
out of place. Additionally other addons aren't beholden to even honor
an empty rank string either and may default (or just entirely omit)
ranks causing the same issue as nameplates.