Make window buttons look prettier in Retail
This updates the maximize, minimize, and resize buttons on the main
frame to use newer art assets in retail clients. The button positioning
has been slightly altered to no longer bleed outside the frame.

On classic clients the old assets are used, as the newer art doesn't
exist there.

For the resize button this doesn't update the styling of
the TRP3_ResizeButton template itself. The new templates use a different
frame size and I didn't want to spend any time going through and
correcting every single usage of the resize button template.