Separate methods from distributions
Would rather have broadcast methods be independent of the actual
distribution type used for the transmission.

The TRP3_BroadcastMethod enumeration that was really only intended
for internal use has been moved to TRP3_API to signal it's now intended
to be used elsewhere more publicly.

The contents of the enum have changed to only include World, Guild, and
Group options. These get mapped to appropriate comms channels by the
broadcast code.

The client feature flag for comms has been altered to just be a boolean
for detecting channel support; as all the code that uses this feature
was specifically just checking "is it a channel" it makes sense to just
have that be a boolean flag.