Add additional checks for group membership
As we're testing if the player is in a guild before sending on that
channel, it makes sense to apply a similar check for group messages.

We'll handle the downlevelling of RAID to PARTY manually as part of
this - it simplifies the checks a bit and also gives us some insurance
in case the underlying API one day stops redirecting RAID to PARTY
for us.