Rework launcher button integration (#768)
* Rework launcher button integration

The "launcher" registered through LibDBIcon has been split off into
a new module and can be disabled by the user if desired via our
module settings page.

For the addon compartment integration we're now using LibDBCompartment
instead of the TOC fields. This provides us with the ability to
dynamically unregister ourselves from the compartment at runtime.

The launcher now has its own settings page with a few additional
options.

  - Lock minimap button
  - Reset minimap button position
  - Show in addon compartment

Finally - the actions undertaken by the launcher when clicked can now
be fully customized. It's possible to for example configure it such
that a middle-click toggles your roleplay status, or for a modified
click to open directly to your current profile.

* Refresh tooltip on binding changes

* Moving launcher locale keys

* Thank you Github very cool

---------

Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>
Co-authored-by: Solanya <solanya@totalrp3.info>