Persist window layout in configuration (#803)
This replaces the use of the game layout cache for storing the position
of our main windowto instead persist it within our own saved variables,
using a combination of LibWindow for window positioning and our own
management of window size.

The layout cache is known to have a few issues - it restores layout
information at weird times in the load process, and it forgets frames
if they're not loaded in a UI session.

Storing the layout in our saved variables resolves this; we get a fully
controllable point in the load process from which we can restore the
layout and reposition/resize the window, and we don't have to worry
about the game forgetting our frame if the addon isn't loaded once.

Use of LibWindow for managing the window position is employed as it's
the recommended practice; it has smart logic for anchoring to the
nearest quadrant of the screen and handles scale changes. This can be
quite important for the small subset of users who might be mirroring
their SVs across devices with different screen resolutions.