Use MaskTextures instead of SetPortraitToTexture (#828)
Fixes #827. The SetPortraitToTexture API only accepts textures whose
size is 64x64 or smaller, whereas MaskTextures support any size.

Previously we guarded this with a pcall check to fall back and show a
failure texture (or an unmasked one) if a texture was larger than 64x64
however a variant of WoWUIBugs#393 impacts this API - the pcall will
return true implying no error occurred, but instead the error about
the texture size will be deferred and raised upon the next call into
any widget API - as seen in the ticket.

This also removes the applyRoundTexture utility in general - this
requires some changes to Extended.