Fix handling of unknown music names
If a profile is received from a player who has selected profile music
from a newer version of LibRPMedia, then the GetMusicNameByFile API
will return nil.

This results in a few errors with the target bar when the player
is targeted (we attempt to format a nil value into a string), and
in the About page we don't display any name at all for the music.

This adds a fallback to the Utils.music.getTitle function so that
in cases where no name can be obtained, we'll instead use a temporary
string that at least shows _something_ and prevents errors.