Don't wait forever on data that's not coming
Currently, a bug exists whereby requests can get stuck if a response to
a prior request for the same data hasn't completed.

For the sake of example, we'll use the term 'sender' for someone who
sends a request, and 'target' for the person who responds to a request.

If a sender initiates an exchange for the target's about page, and
mid-way through receipt the targe reloads their UI, disconnects, or
throws their PC out of the window then the sender keeps a stale entry
around in the 'CURRENT_QUERY_EXCHANGES' table that prevents them from
ever re-requesting the about page until the sender reloads ther UI, logs
out, or throws their own PC out of the window.

This commit resolves that by just turning that table into a cooldown
system; if there's an outstanding request for data then we won't issue a
new request _unless_ an obscenely long amount of time has already
passed. This is currently defined as five minutes.

Additionally to cut down on memory usage a bit, the layout of the
'CURRENT_QUERY_EXCHANGES' table (now 'QueryCooldowns') has been
inverted; this means we only allocate a number of tables based upon
possible information types rather than one table per target player ID.