Fix numeric sequencing of chat links (#851)
When generating a chat link that conflicted with a pre-existing
identifier, we'd attempt to uniquify it by appending an incrementing
number.

The problem however is that if the identifier collided two or more
times, we'd continue appending to the already-appended identifier,
resulting in strings like "My Profile12" if two links for that profile
had previously been generated.

This commit fixes that so that we now suffix links in incrementing order
properly, eg. "My Profile", "My Profile 1", "My Profile 2", and so on.