Add basic progress indicator to character tooltips (#852)
This commit adds a new indicator icon to the tooltip that appears if
there's any outstanding data exchange queries currently in-progress for
the displayed character.

The indicator icon itself shows as the last icon at the bottom-left of
the tooltip, after the glance and unread about page icons.

The icon only shows if we're expecting profile data from the target
player. Notably, this means we *don't* display an icon for the initial
'vernum' exchange. If the target player does not send us any data
packets for ~about a minute, then the exchange is considered to be timed
out and the icon will hide.

There's two reasons for not tracking vernum exchanges. Firstly, we send
that query out a lot with a very small cooldown; this would result in
the icon showing and hiding quite a lot and for very small periods of
time if there's no actual new data to download.

The second reason is MSP addons; those never reply to vernum exchanges
and so the icon ends up staying visible until the query times out.

The bulk of this work is in the actual request tracking mechanism, as
our comms protocol doesn't make it trivial to track the actual progress
of any queries without breaking backwards compatibility. Ideally we'd be
able to track the actual progress of queries, but unfortunately the
stream identifier is embedded as part of the serialized message sent
through Chomp, and in most cases ends up at the very end of each
full message.