Improve handling of hex color inputs
The hex color input field currently doesn't like trailing spaces, which
may occur when someone copy/pastes a color code from a Discord message
where a user reported an issue with this very field.

We now trim the input and reuse the ParseColorFromHexString utility to
remove the need to check the format of the input.