Use sending player's map ID for culling
Currently when a map scan response is received from another player, we
attempt to cull responses if they're sitting in a phased zone while not
part of our group.

The issue is that we're performing this culling based off *our* zone,
not theirs, so scanning a phased map like a garrison has different
results based upon whether or not you are actually in the same map.

We now send the player's map ID as part of map responses and test
against that. If we don't receive a map ID (older client), we'll just
make one up - per the other PR this whole logic has been broken all
along anyway so it doesn't matter much.