Update phased zone constants for garrisons (#866)
The UiMap IDs we have embedded for garrison detection are completely
out-of-touch with reality.

The IDs we presently have are for Telogrus Rift (971), and Tol Dagor
(976). Consequently this means that one cannot launch a map scan while
standing in any of those zones, as we're testing the wrong constants.

Additionally - this phase culling logic is using the players' own zone
rather than the one they actually scanned, but that'll be fixed
separately.