Use displayed map ID rather than sending it back
As pointed out, we obviously know what map we scanned and senders
already do the culling based off the map we request, so we can just base
our response culling off whatever map we're presently looking at.

This also has the benefit of working without an upgrade, but
unfortunately means we no longer have a need to put players in the
ocean.