Patch 10.2.7 compatibility (#862)
* Fix duplicate file load warnings

Probably just ancient merge damage - these three files are loaded in the
same XML file just a few lines up.

* Use new C_StableInfo APIs for pet browser

* Nuke vertical alignment in pet browser

* Fix up additional use of GetStablePetInfo

Unlike the pet browser, this needs to be classic-compatible.

* Cull most GetMouseFocus calls

This is being removed in 11.0.0. One call remains as it's a bit hairy
to migrate and I need to understand what it's actually doing a bit more.

Also need to accomodate Extended which does one similar call to these.

* Migrate to C_AddOns APIs

The non-namespaced variants are deprecated and will be removed in 11.0.
We've held off removing the old usages because Wrath didn't have all the
C_AddOns variants, but this is no longer the case in Cataclysm.

* Migrate to C_Item APIs

Similar story; will be removed in 11.0 and Cataclysm now has these.

* Migrate to C_PvP APIs

Only one usage in this namespace needs migrating. This function is
removed in 11.0, and is present in Cataclysm already.

* Migrate to C_UnitAuras APIs

Non-namespaced UnitAura is dead in 11.0, with the replacement now
available in Cataclysm. This is a bit more complex a migration because
it returns a structured table now, unlike all the others.

Additionally - fixed an improper initialization of the current mount
spell ID in Classic.

* Cull last GetMouseFocus call

As Classic won't get this change for a while, we can fold this into
10.2.7 as having a backward compatible path is forced upon us.

* Remove UIButtonMixin backport

This hasn't been needed since patch 3.4.3 went live.

* No more dimensions

Size and Offset elements have supported inline absolute attributes for a
long time now. Let's do a quick batch cleanup...

* No more insets

These also support inline absolute attributes.