Re-add support for linking companions on Classic
This bonks the code over the head enough times to give it a severe
enough concussion that linking companion pets should hopefully now work
again on Classic.

As part of this, the need to rename a pet has been removed on all
clients. Unrenamed pets in retail will show a warning in their tooltip
however, and get sorted to the list as well as deduplicated.