Adjust clamp insets when mount tooltip is shown
For the strange people who have tooltip on cursor shown, they can run
into issues if the character they're hovering is on a mount and their
tooltip is near the top of the screen - in such a case, the two tooltips
overlap.

This can cheaply be solved by just forcing a larger clamp inset on the
character tooltip when we show (or hide) the mount tooltip.