Add soft target support for tooltips
This implements the necessary changes to detach our tooltips from
strictly looking at the "mouseover" unit and to instead use whatever
unit is currently considered active per the world cursor.

The effect of this is that we now properly customize tooltips when a
user has the SoftTargetFriend and SoftTargetTooltipFriend CVars enabled.
Previously, if these were enabled then the tooltip that appears when
looking at a friendly player would be the standard game one and wouldn't
be customized with RP info.

One other side effect of this support is that the UPDATE_MOUSE_OVER
event can now fire for non-"mouseover" units, which means that we'll
also fire off profile requests if the appropriate soft target CVars are
enabled and the user is looking at another friendly player.