Remove usages of UIDropDownMenuTemplate
Instead, decorate legacy dropdown buttons manually via region creation
in Lua. This means we eliminate any need to ever give dropdown buttons
names, as UIDropDownMenuTemplate has a mixture of "$parentFoo" names on
its subregions.

Downside is that we now need the dropdown button to be 32 units in
height. The reason for this is that UIDD_Initialize is stupid and sets
the height of the dropdown frame - so unless the size matches, it'd
cause the widget to jump.