Port UnitPopup module to new menus
Note that this requires some substantial changes. As part of this,
compatibility with clients that don't have the new menus is entirely
dropped.

This should be a temporary thing ultimately, as eventually the new menu
system will make itself into Classic - but the effort required to
support them both in the context of this module is a bit too much.