Remove use of deprecated PetCanBeRenamed API
With the new stable UI in 10.2.7 the restriction on when players can
rename Hunter pets has been removed. In 11.0, the API is gone
completely.

Instead we'll just detect if a pet is renamable by checking if it can be
abandoned - this matches similar logic used by Blizzard who hide the
"Rename" option in right-click menus for pets that can't be abandoned.